requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt -o requirements.txt --no-dev
attrs==25.1.0 \
    --hash=sha256:1c97078a80c814273a76b2a298a932eb681c87415c11dee0a6921de7f1b02c3e \
    --hash=sha256:c75a69e28a550a7e93789579c22aa26b0f5b83b75dc4e08fe092980051e1090a
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.230 \
    --hash=sha256:9e2281ce1ffe2cdb8d433bd26d3b2c5767eac282871064ab66de9a2ecc987fec \
    --hash=sha256:e41bf095ca74af9924e9b2e3244091ba3298f40b938b2397634f551d6ec8a099
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.1.0 \
    --hash=sha256:1f292c0631f86708ba4ee328b3a2b229f7e46ea1c79fbde567ee9eb119c2b0e2 \
    --hash=sha256:24a388b69a44d03bae6dbf864c4e25ba650d4b61c008b4568b94ffbb9a69e40e
    # via aws-cdk-lib
aws-cdk-cloud-assembly-schema==41.2.0 \
    --hash=sha256:7064ac13f6944fd53f8d8eace611d3c5d8db7014049d629f5c47ede8dc5f2e3b \
    --hash=sha256:779ca7e3edb02695e0a94a1f38e322b04fbe192cd7944553f80b681a21edd670
    # via aws-cdk-lib
aws-cdk-lib==2.188.0 \
    --hash=sha256:0f7c26444eb56dcdcf9b9357a06863859fee51965999717753eaa2af64c9b99f \
    --hash=sha256:373f40217dbb9f924f530cd51cb56b3d952bfe571bfa8702f764f1643cef7e88
    # via
    #   cdk-ecr-deployment
    #   cdk-nag
    #   cloudcomponents-cdk-dynamodb-seeder
    #   s3-cross-region-compressor
cattrs==24.1.3 \
    --hash=sha256:981a6ef05875b5bb0c7fb68885546186d306f10f0f6718fe9b96c226e68821ff \
    --hash=sha256:adf957dddd26840f27ffbd060a6c4dd3b2192c5b7c2c0525ef1bd8131d8a83f5
    # via jsii
cdk-ecr-deployment==4.0.1 \
    --hash=sha256:3faba1a64d341c43f2e8e42a97897c8bfe475bc1f040c3f0fd03f2acf6ccce36 \
    --hash=sha256:f4ebcb6fb11ad9d4ec210eae0e13fe3466b3df1f197ead92bd5783cba999d7e7
    # via s3-cross-region-compressor
cdk-nag==2.36.24 \
    --hash=sha256:699d8e3fe17fd9d2cbf9d51fa9d284a574edf81e1b18bec27fa9a20f18ee551c \
    --hash=sha256:b79196f138386f5ce02f2d94009e5d8baf999d046709f79ad43d0039d4d740cd
    # via s3-cross-region-compressor
certifi==2025.1.31 \
    --hash=sha256:3d5da6925056f6f18f119200434a4780a94263f10d1c21d032a6f6b2baa20651 \
    --hash=sha256:ca78db4565a652026a4db2bcdf68f2fb589ea80d0be70e03929ed730746b84fe
    # via requests
charset-normalizer==3.4.1 \
    --hash=sha256:0924e81d3d5e70f8126529951dac65c1010cdf117bb75eb02dd12339b57749dd \
    --hash=sha256:09b26ae6b1abf0d27570633b2b078a2a20419c99d66fb2823173d73f188ce601 \
    --hash=sha256:22e14b5d70560b8dd51ec22863f370d1e595ac3d024cb8ad7d308b4cd95f8313 \
    --hash=sha256:2967f74ad52c3b98de4c3b32e1a44e32975e008a9cd2a8cc8966d6a5218c5cb2 \
    --hash=sha256:2d074908e1aecee37a7635990b2c6d504cd4766c7bc9fc86d63f9c09af3fa11b \
    --hash=sha256:44251f18cd68a75b56585dd00dae26183e102cd5e0f9f1466e6df5da2ed64ea3 \
    --hash=sha256:44ecbf16649486d4aebafeaa7ec4c9fed8b88101f4dd612dcaf65d5e815f837f \
    --hash=sha256:8436c508b408b82d87dc5f62496973a1805cd46727c34440b0d29d8a2f50a6c9 \
    --hash=sha256:955f8851919303c92343d2f66165294848d57e9bba6cf6e3625485a70a038d11 \
    --hash=sha256:aabfa34badd18f1da5ec1bc2715cadc8dca465868a4e73a0173466b688f29dda \
    --hash=sha256:b1ac5992a838106edb89654e0aebfc24f5848ae2547d22c2c3f66454daa11971 \
    --hash=sha256:c75cb2a3e389853835e84a2d8fb2b81a10645b503eca9bcb98df6b5a43eb8886 \
    --hash=sha256:d98b1668f06378c6dbefec3b92299716b931cd4e6061f3c875a71ced1780ab85 \
    --hash=sha256:eb8178fe3dba6450a3e024e95ac49ed3400e506fd4e9e5c32d30adda88cbd407 \
    --hash=sha256:fa88b843d6e211393a37219e6a1c1df99d35e8fd90446f1118f4216e307e48cd
    # via requests
cloudcomponents-cdk-dynamodb-seeder==2.4.0 \
    --hash=sha256:319f270796d7325858e06fb22b8bd9e53b4226a28aad722668f26751eea2edce \
    --hash=sha256:e81f56713c9715d1ce474bc6b588dd76f27718424bb007ea1ee2d2d4c85e4ca9
    # via s3-cross-region-compressor
constructs==10.4.2 \
    --hash=sha256:1f0f59b004edebfde0f826340698b8c34611f57848139b7954904c61645f13c1 \
    --hash=sha256:ce54724360fffe10bab27d8a081844eb81f5ace7d7c62c84b719c49f164d5307
    # via
    #   aws-cdk-lib
    #   cdk-ecr-deployment
    #   cdk-nag
    #   cloudcomponents-cdk-dynamodb-seeder
idna==3.10 \
    --hash=sha256:12f65c9b470abda6dc35cf8e63cc574b1c52b11df2c86030af0ac09b01b13ea9 \
    --hash=sha256:946d195a0d259cbba61165e88e65941f16e9b36ea6ddb97f00452bae8b1287d3
    # via requests
importlib-resources==6.5.2 \
    --hash=sha256:185f87adef5bcc288449d98fb4fba07cea78bc036455dd44c5fc4a2fe78fed2c \
    --hash=sha256:789cfdc3ed28c78b67a06acb8126751ced69a3d5f79c095a98298cd8a760ccec
    # via jsii
jsii==1.112.0 \
    --hash=sha256:6510c223074d9b206fd0570849a791e4d9ecfff7ffe68428de73870cea9f55a1 \
    --hash=sha256:6b7d19f361c2565b76828ecbe8cbed8b8d6028a82aa98a46b206a4ee5083157e
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   cdk-ecr-deployment
    #   cdk-nag
    #   cloudcomponents-cdk-dynamodb-seeder
    #   constructs
publication==0.0.3 \
    --hash=sha256:0248885351febc11d8a1098d5c8e3ab2dabcf3e8c0c96db1e17ecd12b53afbe6 \
    --hash=sha256:68416a0de76dddcdd2930d1c8ef853a743cc96c82416c4e4d3b5d901c6276dc4
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   cdk-ecr-deployment
    #   cdk-nag
    #   cloudcomponents-cdk-dynamodb-seeder
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0 \
    --hash=sha256:37dd54208da7e1cd875388217d5e00ebd4179249f90fb72437e91a35459a0ad3 \
    --hash=sha256:a8b2bc7bffae282281c8140a97d3aa9c14da0b136dfe83f850eea9a5f7470427
    # via jsii
requests==2.32.4 \
    --hash=sha256:27babd3cda2a6d50b30443204ee89830707d396671944c998b5975b031ac2b2c \
    --hash=sha256:27d0316682c8a29834d3264820024b62a36942083d52caf2f14c0591336d3422
    # via s3-cross-region-compressor
six==1.17.0 \
    --hash=sha256:4721f391ed90541fddacab5acf947aa0d3dc7d27b2e1e8eda2be8970586c3274 \
    --hash=sha256:ff70335d468e7eb6ec65b95b99d3a2836546063f63acc5171de367e834932a81
    # via python-dateutil
typeguard==2.13.3 \
    --hash=sha256:00edaa8da3a133674796cf5ea87d9f4b4c367d77476e185e80251cc13dfbb8c4 \
    --hash=sha256:5e3e3be01e887e7eafae5af63d1f36c849aaa94e3a0112097312aabfa16284f1
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   cdk-ecr-deployment
    #   cdk-nag
    #   cloudcomponents-cdk-dynamodb-seeder
    #   constructs
    #   jsii
typing-extensions==4.13.1 \
    --hash=sha256:4b6cf02909eb5495cfbc3f6e8fd49217e6cc7944e145cdda8caa3734777f9e69 \
    --hash=sha256:98795af00fb9640edec5b8e31fc647597b4691f099ad75f469a2616be1a76dff
    # via jsii
urllib3==2.3.0 \
    --hash=sha256:1cee9ad369867bfdbbb48b7dd50374c0967a0bb7710050facf0dd6911440e3df \
    --hash=sha256:f8c5449b3cf0861679ce7e0503c7b44b5ec981bec0d1d3795a07f1ba96f0204d
    # via requests
